api 함수 추가
