+++ src/remotes/add-favorite-station/index.ts
@@ -1,0 +1,5 @@
+import { http } from "@/utils/http";
+
+export function addFavoriteStation(stationId: number) {
+	return http.post<string>(`/stations/${stationId}/favorite`, {});
+}
