+++ src/remotes/add-favorite-station/mock.ts
@@ -1,0 +1,8 @@
+import { http, HttpResponse } from "msw";
+
+export const addFavoriteStationMock = http.post(
+	"/api/stations/:stationId/favorite",
+	() => {
+		return HttpResponse.json<string>("Station added to favorites");
+	},
+);
