+++ src/remotes/calculate-day-of-week-statistics/index.ts
@@ -1,0 +1,9 @@
+import { http } from "@/utils/http";
+
+export type StatisticsRequest = {
+	requestedAt: string;
+};
+
+export function calculateDayOfWeekStatistics(request: StatisticsRequest) {
+	return http.post<string>("/statistics/day-of-week", request);
+}
