+++ src/remotes/calculate-temperature-statistics/index.ts
@@ -1,0 +1,9 @@
+import { http } from "@/utils/http";
+
+export type StatisticsRequest = {
+	requestedAt: string;
+};
+
+export function calculateTemperatureStatistics(request: StatisticsRequest) {
+	return http.post<string>("/statistics/temperature", request);
+}
