+++ src/remotes/delete-comment/index.ts
@@ -1,0 +1,5 @@
+import { http } from "@/utils/http";
+
+export function deleteComment(commentId: number) {
+	return http.delete<string>(`/posts/comments/${commentId}`);
+}
