+++ src/remotes/delete-favorite-station/index.ts
@@ -1,0 +1,5 @@
+import { http } from "@/utils/http";
+
+export function deleteFavoriteStation(stationId: number) {
+	return http.delete<string>(`/stations/${stationId}/favorite`);
+}
