+++ src/remotes/delete-favorite-station/mock.ts
@@ -1,0 +1,8 @@
+import { http, HttpResponse } from "msw";
+
+export const deleteFavoriteStationMock = http.delete(
+	"/api/stations/:stationId/favorite",
+	() => {
+		return HttpResponse.json<string>("Station removed from favorites");
+	},
+);
