+++ src/remotes/delete-post/index.ts
@@ -1,0 +1,5 @@
+import { http } from "@/utils/http";
+
+export function deletePost(postId: number) {
+	return http.delete<string>(`/posts/${postId}`);
+}
