+++ src/remotes/health-check/index.ts
@@ -1,0 +1,5 @@
+import { http } from "@/utils/http";
+
+export function healthCheck() {
+	return http.get<string>("/health");
+}
