+++ src/remotes/health-check/mock.ts
@@ -1,0 +1,5 @@
+import { http, HttpResponse } from "msw";
+
+export const healthCheckMock = http.get("/api/health", () => {
+	return HttpResponse.json<string>("OK");
+});
