+++ src/remotes/is-favorite-station/index.ts
@@ -1,0 +1,9 @@
+import { http } from "@/utils/http";
+
+export interface IsFavoriteStationResponse {
+	isFavorite: boolean;
+}
+
+export function isFavoriteStation(stationId: number) {
+	return http.get<IsFavoriteStationResponse>(`/stations/${stationId}/favorite`);
+}
