+++ src/remotes/join-board/index.ts
@@ -1,0 +1,5 @@
+import { http } from "@/utils/http";
+
+export function joinBoard(stationId: number) {
+	return http.post<string>(`/stations/${stationId}/posts/join`, {});
+}
