+++ src/remotes/like-post/index.ts
@@ -1,0 +1,10 @@
+import { http } from "@/utils/http";
+
+export interface PostLikeResponse {
+	postId: number;
+	likeCount: number;
+}
+
+export function likePost(postId: number) {
+	return http.post<PostLikeResponse>(`/posts/${postId}/like`, {});
+}
