+++ src/remotes/read-weather-histories/index.ts
@@ -1,0 +1,5 @@
+import { http } from "@/utils/http";
+
+export function readWeatherHistories() {
+	return http.post<string>("/weather/upload", {});
+}
