+++ src/remotes/read-weather-histories/mock.ts
@@ -1,0 +1,5 @@
+import { http, HttpResponse } from "msw";
+
+export const readWeatherHistoriesMock = http.post("/api/weather/upload", () => {
+	return HttpResponse.json<string>("Weather data uploaded successfully");
+});
