+++ src/remotes/unlike-post/index.ts
@@ -1,0 +1,10 @@
+import { http } from "@/utils/http";
+
+export interface PostLikeResponse {
+	postId: number;
+	likeCount: number;
+}
+
+export function unlikePost(postId: number) {
+	return http.delete<PostLikeResponse>(`/posts/${postId}/like`);
+}
